#23 - [REFACTOR] DOMAIN : 도메인 계층의 역할과 책임의 영역을 변경함 그에따른 리팩토링
