+++ apps/server-auth/src/main/java/com/assetmind/server_auth/global/config/SecurityConfig.java
@@ -2,14 +2,14 @@ package com.assetmind.server_auth.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.socket.EnableWebSocketSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 /**
  * 보안 관련 설정
  */
 @Configuration
-@EnableWebSocketSecurity
+@EnableWebSecurity
 public class SecurityConfig {
 
     /**
