+++ apps/server-auth/src/test/java/com/assetmind/server_auth/user/domain/vo/PasswordTest.java
@@ -2,13 +2,12 @@ package com.assetmind.server_auth.user.domain.vo;
 
 import static org.assertj.core.api.Assertions.*;
 
-import com.assetmind.server_auth.user.domain.port.PasswordEncoder;
+import com.assetmind.server_auth.user.application.port.PasswordEncoder;
 import com.assetmind.server_auth.user.infrastructure.security.TestPasswordEncoder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 /**
  * Password VO 객체 단위 테스트
@@ -26,7 +25,7 @@ class PasswordTest {
         String rawPassword = "testPassword1234";
 
         // when
-        Password password = Password.create(rawPassword, encoder);
+        Password password = Password.from(encoder.encode(rawPassword));
 
         // then
         assertThat(password.value()).isEqualTo("ENCODE" + rawPassword); // 암호화 확인
@@ -35,24 +34,11 @@ class PasswordTest {
 
     @ParameterizedTest
     @NullAndEmptySource
-    @ValueSource(strings = {"짧다", "1234567", "이테스트비밀번호는20자초과입니다다다다다"})
-    @DisplayName("실패: 비밀번호가 8자 미만이거나 20자 초과이면 예외가 발생한다.")
+    @DisplayName("실패: 비밀번호가 null이나 빈 값이라면 예외가 발생한다.")
     void givenInvalidRawPassword_whenCreate_thenThrowException(String invalidPassword) {
-        assertThatThrownBy(() -> Password.create(invalidPassword, encoder))
+        assertThatThrownBy(() -> Password.from(invalidPassword))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("비밀번호는 8자 이상");
-    }
-
-    @Test
-    @DisplayName("성공: 입력한 평문이 암호화된 값과 일치하는지 확인한다.")
-    void givenRawAndEncodedPassword_whenMatch_thenReturnBoolean() {
-        // given
-        String raw = "password123";
-        Password password = Password.create(raw, encoder);
-
-        // when & then
-        assertThat(password.match(raw, encoder)).isTrue();
-        assertThat(password.match("wrongPassword", encoder)).isFalse();
+                .hasMessageContaining("비밀번호는 필수");
     }
 
 }
