+++ apps/server-auth/src/test/java/com/assetmind/server_auth/user/infrastructure/security/TestPasswordEncoder.java
@@ -1,6 +1,6 @@
 package com.assetmind.server_auth.user.infrastructure.security;
 
-import com.assetmind.server_auth.user.domain.port.PasswordEncoder;
+import com.assetmind.server_auth.user.application.port.PasswordEncoder;
 
 public class TestPasswordEncoder implements PasswordEncoder {
